requirements.txt
typer[all]>=0.9.0
python-dotenv>=1.0.0
groq>=0.4.0
watchdog>=3.0.0
